+++ src/error.rs
@@ -1,0 +1,61 @@
+use warp::{
+    filters::{
+        body::BodyDeserializeError,
+        cors::CorsForbidden,
+    },
+    reject::Reject,
+    Rejection,
+    Reply,
+    http::StatusCode,
+};
+
+#[derive(Debug)]
+pub enum Error {
+    ParseError(std::num::ParseIntError),
+    MissingParameters,
+    QuestionNotFound,
+}
+
+impl std::fmt::Display for Error {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        // Why *self and not just self (&self)?
+        // Would it then only match on a reference?
+        match *self {
+            Error::ParseError(ref err) => write!(f, "Cannot parse parameter: {}", err),
+            Error::MissingParameters => write!(f, "Missing parameter"),
+            Error::QuestionNotFound => write!(f, "Question not Found"),
+        }
+    }
+}
+
+impl Reject for Error {}
+
+#[derive(Debug)]
+struct InvalidId;
+impl Reject for InvalidId {}
+
+// Any error that doesn't match the first two conditions will return
+// "Route not found". Not ideal.
+pub async fn return_error(r: Rejection) -> Result<impl Reply, Rejection> {
+    if let Some(error) = r.find::<Error>() {
+        Ok(warp::reply::with_status(
+            error.to_string(),
+            StatusCode::RANGE_NOT_SATISFIABLE,
+        ))
+    } else if let Some(error) = r.find::<CorsForbidden>() {
+        Ok(warp::reply::with_status(
+            error.to_string(),
+            StatusCode::FORBIDDEN,
+        ))
+    } else if let Some(error) = r.find::<BodyDeserializeError>() {
+        Ok(warp::reply::with_status(
+            error.to_string(),
+            StatusCode::UNPROCESSABLE_ENTITY,
+        ))
+    } else {
+        Ok(warp::reply::with_status(
+            "Route not found".to_string(),
+            StatusCode::NOT_FOUND,
+        ))
+    }
+}
