+++ src/types/answer.rs
@@ -1,0 +1,13 @@
+use serde::{Serialize, Deserialize};
+
+use crate::types::question::QuestionId;
+
+#[derive(Deserialize, Serialize, Debug, Clone, PartialEq, Eq, Hash)]
+pub struct AnswerId(pub String);
+
+#[derive(Debug, Deserialize, Serialize, Clone)]
+pub struct Answer {
+    pub id: AnswerId,
+    pub content: String,
+    pub question_id: QuestionId,
+}
