+++ src/types/mod.rs
@@ -1,0 +1,3 @@
+pub mod answer;
+pub mod question;
+pub mod pagination;
